use 'op inject' instead of 'op run'
This has a few advantages over using `op run`:

- No need to backup OP_SESSION_* variables. These will not be output
from 'op inject'.
- With `cat` inside the process substitution it broke using ctrl+c to
cancel the process. The shell needed to be killed whenever 1password was
not responding.

A side-advantage: it _can_ handle variable substitution inside of op://
URLs. Not interesting when passing these through bash, but it is
interesting when passing a file that holds these URLs.